requirements.txt
Package             Version
------------------- -------
Django              2.2.1  
django-crispy-forms 1.7.2  
djangorestframework 3.10.3 
pip                 19.2.3 
pytz                2019.1 
setuptools          41.0.1 
sqlparse            0.3.0  
wheel               0.33.2 
